refactor: reorganize component into folders
